Uploaded QA program for 12.01.2023-15.01.2023
The program simulates the working of the coffee machine. User chooses the drink from the list and amount of sugar, and payment method (with change if too much cash was given, and cash paid is stored in the machine). If stock values are too short, the warning will be displayed in the main menu, and drinks with insufficient ingredients won't be able to be ordered. 
There's also a hidden 0 option, protected by password, which accesses the admin menu. In this menu current ingredients and cash in the machine is displayed, and also options to add a new drink, and to restock, returning all values of ingredients (and cash) to initial values.

Main.py imports other classes, and checks password for admin menu
coffee_machine.py realizes process_payment method, dispense method, __init__ method (counts sugar, milk, and coffee), restock and check_restock_warning methods.
menu.py displays drinks from drinks.txt, and processes user choices during operation
admin.py realizes admin options of displaying ingredients and cash inside, of adding new drink to the menu, and replenish stock (returning to starting values, including cash).
drinks.txt is the list of drinks, and required ingredients for their dispensing (excluding sugar, as it's custom selected for every drink by user during operation).